+++ backend/src/main/java/de/kaliburg/morefair/game/round/services/mapper/RoundMapper.java
@@ -49,6 +49,7 @@ public class RoundMapper {
                 .maxVinegarThrown(fairConfig.getMaxVinegarThrown())
                 .build()
         )
+        .number(round.getNumber())
         .build();
   }
 }